// src/components/email-agent-modal/form/fields.enum.ts
export enum FormFieldsEnum {
  /**
   * Email
   */
  name = 'name',
  /**
   * Email
   */
  email = 'email',
  /**
   * Phone
   */
  phone = 'phone',
  /**
   * Message
   */
  message = 'message',
  /**
   * Email alert
   */
  emailAlert = 'emailAlert',
  /**
   * Captcha
   */
  captchaToken = 'captcha_token',
  /**
   * Accept terms
   */
  acceptTerms = 'acceptTerms',
  /**
   * Receive advertising
   */
  receiveAdvertising = 'receiveAdvertising',
}